// apps/web/src/lib/generated/models/RegulatoryDesignationType.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RegulatoryDesignationType = 'breakthrough_therapy' | 'fast_track' | 'priority_review' | 'accelerated_approval' | 'orphan_drug' | 'prime' | 'sakigake' | 'conditional_marketing_authorisation' | 'other';