// apps/web/src/lib/generated/models/RegulatoryLabelChangeType.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RegulatoryLabelChangeType = 'initial_label' | 'indication_expansion' | 'population_expansion' | 'restriction' | 'dosing_update' | 'administration_update' | 'safety_update' | 'boxed_warning' | 'contraindication' | 'other';