// apps/web/src/lib/generated/models/RegulatorySafetySignalType.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RegulatorySafetySignalType = 'adverse_event' | 'boxed_warning' | 'contraindication' | 'risk_management' | 'recall' | 'clinical_hold' | 'postmarketing_requirement' | 'other';